requirements.txt
# FastApi
fastapi
fastapi-pagination
starlette
uvicorn
pydantic>=2.4
pydantic-settings
python-dotenv

# Database
tortoise-orm
asyncpg
aerich

# Security
passlib
bcrypt

# Web Requests
httpx

# HTML packages
jinja2
bs4

# Parsing
unidecode

# utility
validators
Faker
